add registration
