+++ src/main/java/stanfordnlp/StanfordCoreNlpProviderServiceApplication.java
@@ -3,7 +3,7 @@ package stanfordnlp;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication //(exclude = { SecurityAutoConfiguration.class })
+@SpringBootApplication
 public class StanfordCoreNlpProviderServiceApplication {
 
 	public static void main(String[] args) {
