+++ src/main/java/stanfordnlp/authentification/RegistrationRequest.java
@@ -1,0 +1,23 @@
+package stanfordnlp.authentification;
+
+public class RegistrationRequest {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
+
