+++ src/main/java/stanfordnlp/authentification/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package stanfordnlp;
+package stanfordnlp.authentification;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -13,12 +13,11 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.provisioning.JdbcUserDetailsManager;
+import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
 import javax.sql.DataSource;
-//@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true, jsr250Enabled = true)
 
 @Configuration
 @EnableWebSecurity
@@ -33,7 +32,7 @@ public class SecurityConfiguration {
                 .disable()
                 .authorizeHttpRequests()
                 .requestMatchers("/health").permitAll()
-//                .antMatchers("/user/**").hasAnyRole("USER", "ADMIN")
+                .requestMatchers("/registration").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .httpBasic()
@@ -44,16 +43,6 @@ public class SecurityConfiguration {
         return http.build();
     }
 
-    //    @Bean
-//    public CustomUserDetailsService userDetailsService(PasswordEncoder passwordEncoder) {
-//        UserDetails user = User.withUsername("user")
-//                .password(passwordEncoder.encode("password"))
-//                .roles("USER")
-//                .build();
-//        CustomUserDetailsService customUserDetailsService = new CustomUserDetailsService();
-//        customUserDetailsService.addUser(user);
-//        return customUserDetailsService;
-//    }
     @Bean
     public AuthenticationManager authenticationManager(HttpSecurity http, PasswordEncoder bCryptPasswordEncoder, UserDetailsService userDetailsService)
             throws Exception {
@@ -64,21 +53,19 @@ public class SecurityConfiguration {
                 .build();
     }
 
-
-
     @Bean
-    public UserDetailsService userDetailsService(PasswordEncoder bCryptPasswordEncoder) {
+    public UserDetailsManager userDetailsService(PasswordEncoder bCryptPasswordEncoder) {
         JdbcUserDetailsManager manager = new JdbcUserDetailsManager();
-//        InMemoryUserDetailsManager manager = new InMemoryUserDetailsManager();
         manager.setDataSource(datasource);
-        manager.createUser(User.withUsername("user")
-                .password(bCryptPasswordEncoder.encode("password"))
-                .roles("USER")
-                .build());
+//        manager.createUser(User.withUsername("user")
+//                .password(bCryptPasswordEncoder.encode("password"))
+//                .roles("USER")
+//                .build());
         return manager;
     }
+
     @Bean
-    public PasswordEncoder passwordEncoder() {
+    public BCryptPasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
