+++ src/main/resources/application.properties
@@ -3,9 +3,4 @@ spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.username=sa
 spring.datasource.password=default
 spring.datasource.driver-class-name=org.h2.Driver
-#
-##spring.security.user.name=user
-##spring.security.user.password=password
-#
-# Enable H2 Console (optional, for database administration)
 spring.h2.console.enabled=true
